fix cos, pxe1->pxe2

+++ src/p32e2.rs
@@ -225,19 +225,15 @@ impl P32E2 {
     }
 
     #[inline]
-    pub(crate) fn calculate_regime(k: i8) -> (u32, bool, u8) {
+    pub(crate) fn calculate_regime(k: i8) -> (u32, bool, u32) {
         let reg;
         if k < 0 {
-            reg = (-k) as u8;
-            (
-                0x4000_0000_u32.checked_shr(reg as u32).unwrap_or(0),
-                false,
-                reg,
-            )
+            reg = (-k) as u32;
+            (0x4000_0000_u32.checked_shr(reg).unwrap_or(0), false, reg)
         } else {
-            reg = (k + 1) as u8;
+            reg = (k + 1) as u32;
             (
-                0x7fff_ffff - 0x7fff_ffff_u32.checked_shr(reg as u32).unwrap_or(0),
+                0x7fff_ffff - 0x7fff_ffff_u32.checked_shr(reg).unwrap_or(0),
                 true,
                 reg,
             )
